Up#12
Update cart

+++ src/components/cart/cart.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import { useNavigate, useParams} from "react-router-dom";
+import { useNavigate, useParams } from "react-router-dom";
 import axios from "axios";
 import "../cart/cart.css";
 import cat from "../../img/bonludan.jpeg";
@@ -84,19 +84,39 @@ export default function Cart() {
 
   const handleDeleteItem = async (productid) => {
     try {
-      const response = await axios.delete(
-        `http://localhost:3001/v1/cart/deleteItemById/${customerid}`,
-        {
-          data: { productid: productid._id },
-        }
-      );
+      if (productid.age) {
+        const response = await axios.delete(
+          `http://localhost:3001/v1/cart/deleteItemById/${customerid}`,
+          {
+            data: { petid: productid._id },
+          }
 
-      if (response.status === 200) {
-        console.log("Sản phẩm đã được xóa thành công");
-        updateCartData();
-      } else {
-        console.error("Lỗi khi xóa sản phẩm:", response.data.message);
+        );
+        if (response.status === 200) {
+          console.log("Sản phẩm đã được xóa thành công");
+          updateCartData();
+        } else {
+          console.error("Lỗi khi xóa sản phẩm:", response.data.message);
+        }
       }
+      if (productid.quantity) {
+        const response = await axios.delete(
+          `http://localhost:3001/v1/cart/deleteItemById/${customerid}`,
+          {
+            data: { productid: productid._id },
+          }
+
+        );
+        if (response.status === 200) {
+          console.log("Sản phẩm đã được xóa thành công");
+          updateCartData();
+        } else {
+          console.error("Lỗi khi xóa sản phẩm:", response.data.message);
+        }
+      }
+
+
+
     } catch (error) {
       console.error("Lỗi khi xóa sản phẩm:", error);
     }
@@ -109,15 +129,17 @@ export default function Cart() {
       );
       const newData = response.data.cart;
       const newProduct = response.data.cart.products;
+      const newPet = response.data.cart.pets;
 
 
       const initialSelectedProducts = newProduct.reduce((acc, product) => {
         return { ...acc, [product.productid._id]: false };
       }, {});
 
+
       setCart(newData);
       setProduct(newProduct);
-      
+      setPet(newPet);
       setSelectedProducts(initialSelectedProducts);
 
       localStorage.setItem("Cart", JSON.stringify(newData));
@@ -169,7 +191,7 @@ export default function Cart() {
     for (const product of products) {
       const productId = product.productid._id;
       if (selectedProducts[productId]) {
-        productsQ ++;
+        productsQ++;
       }
     }
